Additions
Program works again. Private Notes are now ecrypted. A lot more of changes.

+++ app/database/delete.php
@@ -5,7 +5,17 @@
 // Function that deletes a public note by the note's ID.
 function delete_public_note($note_id)
 {
-    $conn = create_connection_and_return_conn_variable();
+    require "../../memory.php";
+    $conn = new mysqli(
+        $_ENV['database_server_name'],
+        $_ENV['database_username'],
+        $_ENV['database_password'],
+        $_ENV['database_name']
+    );
+
+    if ($conn->connect_error) {
+        die("Database connection failed: " . $conn->connect_error);
+    }
 
     if ($stmt = $conn->prepare("DELETE FROM NOTE WHERE ID = ? AND owner_id = 'public';")) {
         $stmt->bind_param("i", $note_id);
@@ -24,12 +34,21 @@ function delete_public_note($note_id)
 // Function that deletes a private note.
 function delete_private_note($note_id)
 {
-    require '../../../config.php';
-    require '../../../memory.php';
+    require "../../memory.php";
 
     // If user is not logged in, they will be automatically redirected to project root route.
     if (isset($_SESSION['user_logged_in'])) {
-        $conn = create_connection_and_return_conn_variable();
+        require "../../memory.php";
+        $conn = new mysqli(
+            $_ENV['database_server_name'],
+            $_ENV['database_username'],
+            $_ENV['database_password'],
+            $_ENV['database_name']
+        );
+    
+        if ($conn->connect_error) {
+            die("Database connection failed: " . $conn->connect_error);
+        }
 
         // The note is deleted by their associated owner's ID. This ID is stored inside 
         // SESSION variables. This prevents further errors inside the database.
@@ -52,8 +71,17 @@ function delete_private_note($note_id)
 // Function that deletes an user account stored into the database.
 function delete_user($user_id)
 {
-    require '../../config.php';
-    $conn = create_connection_and_return_conn_variable();
+    require "../memory.php";
+    $conn = new mysqli(
+        $_ENV['database_server_name'],
+        $_ENV['database_username'],
+        $_ENV['database_password'],
+        $_ENV['database_name']
+    );
+
+    if ($conn->connect_error) {
+        die("Database connection failed: " . $conn->connect_error);
+    }
 
     if ($stmt = $conn->prepare("DELETE FROM USER WHERE ID = ?")) {
         $stmt->bind_param("s", $user_id);
@@ -72,8 +100,17 @@ function delete_user($user_id)
 // ownership.
 function delete_associated_notes($user_id)
 {
-    require '../../config.php';
-    $conn = create_connection_and_return_conn_variable();
+    require "../memory.php";
+    $conn = new mysqli(
+        $_ENV['database_server_name'],
+        $_ENV['database_username'],
+        $_ENV['database_password'],
+        $_ENV['database_name']
+    );
+
+    if ($conn->connect_error) {
+        die("Database connection failed: " . $conn->connect_error);
+    }
 
     if ($stmt = $conn->prepare("DELETE FROM NOTE WHERE owner_id = ?")) {
         $stmt->bind_param("s", $user_id);
